fix: ai suggest
